Order Integration Tests Refactor

+++ src/main/java/com/jcanseco/inventoryapi/dtos/orders/OrderDetailsDto.java
@@ -1,11 +1,7 @@
 package com.jcanseco.inventoryapi.dtos.orders;
 
 import com.jcanseco.inventoryapi.dtos.customers.CustomerDto;
-import com.jcanseco.inventoryapi.dtos.purchases.PurchaseItemDto;
-import com.jcanseco.inventoryapi.dtos.suppliers.SupplierDto;
 import lombok.*;
-
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -18,7 +14,7 @@ import java.util.List;
 public class OrderDetailsDto {
     private Long id;
     private CustomerDto customer;
-    private BigDecimal total;
+    private Double total;
     private List<OrderItemDto> items;
     private boolean delivered;
     private LocalDateTime orderedAt;
