+++ src/main/java/com/jcanseco/inventoryapi/dtos/orders/OrderDto.java
@@ -1,8 +1,6 @@
 package com.jcanseco.inventoryapi.dtos.orders;
 
 import lombok.*;
-
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @EqualsAndHashCode
@@ -14,7 +12,7 @@ import java.time.LocalDateTime;
 public class OrderDto {
     private Long id;
     private String customer;
-    private BigDecimal total;
+    private Double total;
     private boolean delivered;
     private LocalDateTime orderedAt;
     private LocalDateTime deliveredAt;
